Chat backend bot/requirements.txt
flask
openai
pypdf2
pyjwt
tiktoken
python-dotenv
